dpkg: On Linux finish writeback before fsync
The second sync_file_range() call, with the operation
SYNC_FILE_RANGE_WAIT_BEFORE, will block until the previously
initiated writeback has completed.

We finish the writeback for all files before calling fsync. This
basically ensures that the delayed allocation has been resolved;
that is, the data blocks have been allocated and written, and the
inode updated (in memory), but not necessarily pushed out to disk.
So that later fsync can become no-ops, minimizing the number of
(costly) jbd2 commits.

Suggested-by: Ted Ts'o <tytso@mit.edu>
Based-on-patch-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Guillem Jover <guillem@debian.org>